submissions/sangampaudel530/requirements.txt
numpy
pandas
opencv-python
matplotlib
scikit-image
rasterio
shapely
geopandas
torch
torchvision
tqdm
Pillow
